# [ready] perf: canonicalize less if not needed