@tinyb0t 2023-08-26
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/ops.py                            1               0
tinygrad/tensor.py                        20              20
------------------------------------------------------------
total                                     21              20
------------------------------------------------------------
lines added in the tinygrad folder: 1
```