@wozeparrot 2023-08-26
hmm, how much does just caching give for performance? having to set the flag everywhere is kinda annoying