@roelofvandijk 2023-08-26
I agree, it doesn't look nice. Knowing when a Tensor gets created by a Tensor might be valuable in other ways. Alternatively, `device` could be either a `str` (not parsed)  or a `tuple` if parsed. This could also address the repeated self._device_extra_args() calls in LazyBuffer.

Here's the breakdown, I think the current state is the sweet spot.



```


Only caching


Total time: 0.625214 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 41

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    41                                             @profile
    42                                             def __init__(self, data:Union[int, float, list, LazyBuffer, np.ndarray], device:Optional[str]=None, dtype:Optional[DType]=None, requires_grad:Optional[bool]=None):
    43    180222      44597.5      0.2      7.1      assert dtype is None or isinstance(dtype, DType), f"invalid dtype {dtype}"
    44    180222     130648.9      0.7     20.9      device = Device.canonicalize(device)




Caching + None check before canonicalize


Total time: 0.605264 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 41

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    41                                             @profile
    42                                             def __init__(self, data:Union[int, float, list, LazyBuffer, np.ndarray], device:Optional[str]=None, dtype:Optional[DType]=None, requires_grad:Optional[bool]=None):
    43    180222      50281.9      0.3      8.3      assert dtype is None or isinstance(dtype, DType), f"invalid dtype {dtype}"
    44    180222     125996.7      0.7     20.8      device = Device.DEFAULT if device is None else Device.canonicalize(device)


Caching + None check before canonicalize + disable canonicalization for function call Tensor init


Total time: 0.510681 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 41

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    41                                             @profile
    42                                             def __init__(self, data:Union[int, float, list, LazyBuffer, np.ndarray], device:Optional[str]=None, dtype:Optional[DType]=None, requires_grad:Optional[bool]=None, canonicalize=True):
    43    180222      43105.9      0.2      8.4      assert dtype is None or isinstance(dtype, DType), f"invalid dtype {dtype}"
    44    180222      39866.5      0.2      7.8      if canonicalize: device = Device.DEFAULT if device is None else Device.canonicalize(device)


```