@wozeparrot 2023-08-26
yea this is a much cleaner change. Should the flag be `canonicalize` or something else? like `internal`?

i do kinda like the tuple thing for device, because something else that's needed is a way to pass more than one flag to the runtime, I kinda did it in `ops_shm` by splitting on something else, but we need a way to pass a read-only flag to `ops_disk` and the way i did it for `ops_shm` doesn't really work cuz people can put commas in filenames.

but the one big downside to this is that selecting a device becomes more complicated.