@roelofvandijk 2023-08-26
I wanted to be explicit, but I dislike having one flag to tell the function what to do with another argument. `internal` would be nicer, but I see no reason for it now (because there is no other use yet). A third option would be to explicitly have a `Tensor.create_from` option so.

Used the `tuple` now, `.device` remains a string (for now, otherwise the changes are way larger in terms of comparing device etc.