@wozeparrot 2023-08-28
I think it might be better to have a list of canonicals to check against, because we should know all of them beforehand, excluding extra device args.