@roelofvandijk 2023-08-28
Can do, will be slower though. 

I realized that the `.device` can be set after initialization, and that won't necessarily be canonicalized. See `toCPU`, currently also fragile. This work won't fix that issue.

Could Tensors be immutable? And then new ones get created on change?