@geohot 2023-08-29
More complexity + a subtle bug reintroduced when DEFAULT changes