@roelofvandijk 2023-08-29
@geohot: This was a work in progress, @wozeparrot and I were discussing whether to check against a known list of canonical devices. Canonicalization is 50% of the tensor init time, and there are a lot of tensors.

Should I continue in that direction, or drop this?