@roelofvandijk 2023-08-31
@wozeparrot 

Total time: 0.921717 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 41

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    41                                             @profile
    42                                             def __init__(self, data:Union[int, float, list, LazyBuffer, np.ndarray], device:Optional[str]=None, dtype:Optional[DType]=None, requires_grad:Optional[bool]=None):
    43    184380      46361.8      0.3      5.0      assert dtype is None or isinstance(dtype, DType), f"invalid dtype {dtype}"
    44    184380     449786.3      2.4     48.8      device = Device.canonicalize(device)


Total time: 0.571723 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 41

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    41                                             @profile
    42                                             def __init__(self, data:Union[int, float, list, LazyBuffer, np.ndarray], device:Optional[str]=None, dtype:Optional[DType]=None, requires_grad:Optional[bool]=None):
    43    184380      44427.1      0.2      7.8      assert dtype is None or isinstance(dtype, DType), f"invalid dtype {dtype}"
    44    184380      98161.9      0.5     17.2      device = Device.DEFAULT if device is None else device if device in Device.canonicals else Device.canonicalize(device)